+++ commands/index.js
@@ -352,36 +352,7 @@ ${
   await ctx.reply(statusText, { parse_mode: "Markdown" });
 }
 
-// Создание персонажа
-async function handleCreateCharacter(ctx) {
-  log(`[Commands] Вызвана команда создания персонажа от ${ctx.from.id}`);
 
-  await User.findOrCreate(ctx.from);
-
-  // Если это callback от кнопки, отвечаем на него
-  if (ctx.callbackQuery) {
-    await ctx.answerCbQuery();
-  }
-
-  await characterCreation.startCreation(ctx);
-}
-
-// Показать персонажа
-async function handleShowCharacter(ctx) {
-  const userId = ctx.from.id;
-  const chatId = ctx.chat.id;
-
-  const character = await Character.findActive(userId, chatId);
-
-  if (!character) {
-    await ctx.reply(
-      "❌ У вас нет персонажа!\n\nИспользуйте /create для создания.",
-      { parse_mode: "Markdown" }
-    );
-    return;
-  }
-
-  let display = await character.getFullDisplay();
 
   // Добавляем информацию о смерти
   if (character.hp_current <= 0) {
@@ -394,167 +365,7 @@ async function handleShowCharacter(ctx) {
   await ctx.reply(display, { parse_mode: "Markdown" });
 }
 
-async function handleQuestRoll(ctx) {
-  const userId = ctx.from.id;
-  const chatId = ctx.chat.id;
 
-  try {
-    await ctx.answerCbQuery("🎲 Бросаем кубик...");
-  } catch (error) {
-    // Игнорируем ошибку answerCbQuery
-  }
-
-  try {
-    // Получаем персонажа
-    const character = await Character.findActive(userId, chatId);
-    if (!character) {
-      await ctx.editMessageText("❌ Персонаж не найден!");
-      return;
-    }
-
-    // Проверяем квест
-    const quest = await questSystem.getActiveQuest(chatId);
-    if (!quest) {
-      await ctx.editMessageText("❌ Квест уже завершен или истек!");
-      return;
-    }
-
-    // Бросаем кубик
-    const roll = Math.floor(Math.random() * 20) + 1;
-
-    // Анимация броска
-    const diceEmojis = ["⚀", "⚁", "⚂", "⚃", "⚄", "⚅"];
-    let lastDice = null;
-
-    for (let i = 0; i < 3; i++) {
-      let randomDice;
-      // Гарантируем, что выбираем другой эмодзи
-      do {
-        randomDice = diceEmojis[Math.floor(Math.random() * diceEmojis.length)];
-      } while (randomDice === lastDice && diceEmojis.length > 1);
-
-      lastDice = randomDice;
-
-      try {
-        await ctx.editMessageText(`${randomDice} Бросаем кубик...`);
-      } catch (error) {
-        // Если ошибка из-за одинакового текста, просто пропускаем
-        if (!error.message.includes("message is not modified")) {
-          throw error;
-        }
-      }
-
-      await new Promise((resolve) => setTimeout(resolve, 500));
-    }
-
-    // Выполняем квест
-    const result = await questSystem.executeQuest(character, roll);
-
-    if (!result.success) {
-      await ctx.editMessageText(`❌ ${result.message}`);
-      return;
-    }
-
-    // Формируем сообщение с результатом
-    const config = require("../config/config");
-    const statConfig = config.STATS[result.statUsed];
-    const criticalText =
-      roll === 20
-        ? "⚡ КРИТИЧЕСКИЙ УСПЕХ! ⚡\n"
-        : roll === 1
-        ? "💀 КРИТИЧЕСКИЙ ПРОВАЛ! 💀\n"
-        : "";
-
-    let message = `🎯 **${result.questTitle}**\n\n`;
-    message += criticalText;
-    message += `🎲 Бросок: **${roll}**\n`;
-    message += `${statConfig.emoji} Модификатор ${statConfig.name}: ${
-      result.statModifier >= 0 ? "+" : ""
-    }${result.statModifier}\n`;
-    message += `📊 Итого: **${result.totalRoll}**\n\n`;
-
-    message += `${result.questResult.result_text}\n\n`;
-
-    message += `**Результаты:**\n`;
-    if (result.xpGained > 0) {
-      message += `✨ Опыт: +${result.xpGained} XP\n`;
-    }
-    if (result.xpGained > 0) {
-      message += `✨ Опыт: +${result.xpGained} XP\n`;
-    }
-
-    // Показываем золото отдельно
-    if (result.goldGained > 0) {
-      if (result.lootGold > 0) {
-        message += `💰 Золото: +${
-          result.goldGained - result.lootGold
-        } (квест) +${result.lootGold} (добыча) = **+${result.goldGained}**\n`;
-      } else {
-        message += `💰 Золото: +${result.goldGained}\n`;
-      }
-    } else if (result.goldGained < 0) {
-      message += `💸 Потеряно золота: ${Math.abs(result.goldGained)}\n`;
-    }
-
-    // Показываем найденные предметы
-    if (result.lootItems && result.lootItems.length > 0) {
-      message += `\n🎁 **Найденные предметы:**\n`;
-
-      const rarityEmojis = {
-        common: "⚪",
-        uncommon: "🟢",
-        rare: "🔵",
-        epic: "🟣",
-        legendary: "🟠",
-      };
-
-      for (const item of result.lootItems) {
-        const emoji = rarityEmojis[item.rarity] || "❓";
-        message += `${emoji} ${item.name}\n`;
-      }
-    }
-
-    if (result.damageDealt > 0) {
-      message += `💔 Урон: -${result.damageDealt} HP (${result.characterHp}/${result.characterMaxHp})\n`;
-    }
-
-    if (result.levelUp) {
-      message += `\n🎉 **НОВЫЙ УРОВЕНЬ! ${result.levelUp.from} → ${result.levelUp.to}**\n`;
-
-      if (result.levelUp.abilityPointsGained > 0) {
-        message += `💎 **Получено ${result.levelUp.abilityPointsGained} очка улучшения!** Используйте /improve\n`;
-      }
-    }
-
-    // Проверяем смерть персонажа
-    if (result.isDead) {
-      message += `\n☠️ **ПЕРСОНАЖ ПОГИБ!** ☠️\n`;
-      message += `\n_${character.name} пал смертью храбрых, выполняя опасный квест._\n`;
-      message += `_Покойся с миром, отважный ${character.getClassInfo()}._\n\n`;
-      message += `Используйте /create для создания нового персонажа.`;
-    }
-    await ctx.editMessageText(message, { parse_mode: "Markdown" });
-
-    log(
-      `${character.name} выполнил квест "${result.questTitle}" с результатом ${result.totalRoll}`
-    );
-  } catch (error) {
-    log(`Ошибка выполнения квеста: ${error.message}`, "error");
-
-    try {
-      await ctx.editMessageText(
-        "❌ Произошла ошибка при выполнении квеста.\nПопробуйте еще раз через несколько секунд.",
-        { parse_mode: "Markdown" }
-      );
-    } catch (editError) {
-      // Если не можем отредактировать сообщение, отправляем новое
-      await ctx.reply(
-        "❌ Произошла ошибка при выполнении квеста.\nПопробуйте еще раз через несколько секунд.",
-        { parse_mode: "Markdown" }
-      );
-    }
-  }
-}
 
 async function handleEquipMenu(ctx) {
   const userId = ctx.from.id;
@@ -1683,60 +1494,7 @@ ${quest.description}
   log(`Квест "${quest.title}" выдан вручную для чата ${chatId}`);
 }
 
-// Показать детальную статистику
-async function handleShowStats(ctx) {
-  const userId = ctx.from.id;
-  const chatId = ctx.chat.id;
 
-  const character = await Character.findActive(userId, chatId);
-
-  if (!character) {
-    await ctx.reply(
-      "❌ У вас нет персонажа!\n\nИспользуйте /create для создания.",
-      { parse_mode: "Markdown" }
-    );
-    return;
-  }
-
-  const classConfig = require("../config/config").CLASSES[character.class];
-  const raceConfig = require("../config/config").RACES[character.race];
-
-  let statsText = `📊 **Детальная статистика**\n\n`;
-  statsText += `🎭 ${character.name}\n`;
-  statsText += `${character.getFullTitle()} • ${character.level} уровень\n\n`;
-
-  // Боевые характеристики
-  statsText += `**⚔️ Боевые параметры:**\n`;
-  statsText += `❤️ Здоровье: ${character.hp_current}/${character.hp_max}\n`;
-  statsText += `🎯 Бонус мастерства: +${character.getProficiencyBonus()}\n`;
-  statsText += `🗡️ Основная характеристика: ${classConfig.primaryStat}\n\n`;
-
-  // Все характеристики с бонусами к броскам
-  statsText += `**🎲 Модификаторы бросков:**\n`;
-  const config = require("../config/config");
-  for (const [stat, info] of Object.entries(config.STATS)) {
-    const bonus = character.getRollBonus(stat);
-    const isPrimary = classConfig.primaryStat === stat;
-    statsText += `${info.emoji} ${info.name}: ${bonus >= 0 ? "+" : ""}${bonus}`;
-    if (isPrimary) statsText += " ⭐";
-    statsText += "\n";
-  }
-
-  // Прогресс
-  statsText += `\n**📈 Прогресс:**\n`;
-  statsText += `✨ Опыт: ${character.experience}\n`;
-  statsText += `💰 Золото: ${character.gold}\n`;
-
-  // Расовые особенности
-  if (raceConfig.abilities.length > 0) {
-    statsText += `\n**${raceConfig.emoji} Расовые способности:**\n`;
-    for (const ability of raceConfig.abilities) {
-      statsText += `• ${ability}\n`;
-    }
-  }
-
-  await ctx.reply(statsText, { parse_mode: "Markdown" });
-}
 
 // Показать инвентарь
 async function handleShowInventory(ctx) {
@@ -3289,192 +3047,8 @@ async function handleGraveyard(ctx) {
   }
 }
 
-// Удалить персонажа
-async function handleDeleteCharacter(ctx) {
-  const userId = ctx.from.id;
-  const chatId = ctx.chat.id;
 
-  const character = await Character.findActive(userId, chatId);
-
-  if (!character) {
-    await ctx.reply("❌ У вас нет персонажа!", { parse_mode: "Markdown" });
-    return;
-  }
-
-  await ctx.reply(
-    `⚠️ **Вы уверены, что хотите удалить персонажа?**\n\n` +
-      `Персонаж: ${character.name} (${character.level} уровень)\n` +
-      `Класс: ${character.getClassInfo()}\n\n` +
-      `Это действие необратимо!`,
-    {
-      parse_mode: "Markdown",
-      reply_markup: {
-        inline_keyboard: [
-          [
-            { text: "❌ Да, удалить", callback_data: "delete_confirm" },
-            { text: "✅ Отмена", callback_data: "delete_cancel" },
-          ],
-        ],
-      },
-    }
-  );
-}
-
-// Подтверждение удаления
-async function confirmDeleteCharacter(ctx) {
-  const userId = ctx.from.id;
-  const chatId = ctx.chat.id;
-
-  log(
-    `[Delete] Начинаем удаление персонажа для user ${userId} в чате ${chatId}`
-  );
-
-  const character = await Character.findActive(userId, chatId);
-
-  if (!character) {
-    await ctx.answerCbQuery("Персонаж не найден");
-    log(
-      `[Delete] Персонаж не найден для user ${userId} в чате ${chatId}`,
-      "warning"
-    );
-    return;
-  }
-
-  const characterName = character.name;
-  const characterId = character.id;
-  const characterUserId = character.user_id;
-
-  log(`[Delete] Удаляем персонажа: ${characterName} (ID: ${characterId})`);
-
-  await character.delete();
-
-  // Добавляем небольшую задержку для гарантии записи в БД
-  await new Promise((resolve) => setTimeout(resolve, 100));
-
-  // Проверяем, что персонаж действительно удален
-  const checkDeleted = await Character.findActive(userId, chatId);
-  if (checkDeleted) {
-    log(`[Delete] ОШИБКА: Персонаж ${characterName} не был удален!`, "error");
-
-    // Пробуем еще раз принудительно
-    const db = require("../database");
-    await db.run(
-      "UPDATE characters SET is_active = 0 WHERE id = ? AND user_id = ? AND chat_id = ?",
-      [characterId, characterUserId, chatId]
-    );
-  } else {
-    log(`[Delete] Персонаж ${characterName} успешно удален`);
-  }
-
-  await ctx.answerCbQuery("Персонаж удален!");
-  await ctx.editMessageText(
-    `✅ Персонаж ${characterName} был удален.\n\n` +
-      `Используйте /create для создания нового персонажа.`,
-    { parse_mode: "Markdown" }
-  );
-}
-
-async function handleShowQuest(ctx) {
-  const chatId = ctx.chat.id;
-  const userId = ctx.from.id;
-
-  // Получаем персонажа
-  const character = await Character.findActive(userId, chatId);
-  if (!character) {
-    await ctx.reply(
-      "❌ У вас нет персонажа!\n\nИспользуйте /create для создания.",
-      { parse_mode: "Markdown" }
-    );
-    return;
-  }
-
-  // Проверяем, не мертв ли персонаж
-  if (character.hp_current <= 0) {
-    await ctx.reply(
-      `☠️ **${character.name} мертв!**\n\n` +
-        `Ваш персонаж погиб с честью.\n` +
-        `HP: ${character.hp_current}/${character.hp_max}\n\n` +
-        `Используйте /create для создания нового персонажа.`,
-      { parse_mode: "Markdown" }
-    );
-    return;
-  }
-
-  // Получаем активный квест
-  const quest = await questSystem.getActiveQuest(chatId);
-  if (!quest) {
-    // Проверяем, можем ли получить новый квест
-    const canReceive = await questSystem.canReceiveQuest(chatId);
-
-    if (canReceive.can) {
-      // Пытаемся выдать новый квест
-      const newQuest = await questSystem.assignQuest(chatId);
-      if (newQuest) {
-        await showQuestInfo(ctx, newQuest, character);
-        return;
-      }
-    }
-
-    await ctx.reply(
-      `❌ Нет активного квеста!\n\n${
-        canReceive.reason || "Ждите автоматической выдачи."
-      }`,
-      { parse_mode: "Markdown" }
-    );
-    return;
-  }
-
-  // Показываем информацию о квесте
-  await showQuestInfo(ctx, quest, character);
-}
-
-// Показать информацию о квесте с кнопкой броска
-async function showQuestInfo(ctx, quest, character) {
-  const config = require("../config/config");
-  const statConfig = config.STATS[quest.stat_check];
-  const timeLeft = Math.ceil(
-    (new Date(quest.expires_at) - new Date()) / 1000 / 60
-  );
-
-  const difficultyEmoji = {
-    easy: "🟢",
-    medium: "🟡",
-    hard: "🔴",
-    epic: "🟣",
-    legendary: "⭐",
-  };
-
-  // Получаем модификатор персонажа
-  const statModifier = character.getRollBonus(quest.stat_check);
-  const modSign = statModifier >= 0 ? "+" : "";
-
-  const message = `
-🎯 **АКТИВНЫЙ КВЕСТ**
-
-${difficultyEmoji[quest.difficulty] || "❓"} **${quest.title}**
-${quest.description}
-
-📊 **Проверка:** ${statConfig.emoji} ${statConfig.name}
-🎲 **Ваш модификатор:** ${modSign}${statModifier}
-⏰ **Осталось времени:** ${timeLeft} мин
-💰 **Базовая награда:** ${quest.xp_reward} XP, ${quest.gold_reward} золота
-
-**Как это работает:**
-• Вы бросите 1d20 ${modSign}${statModifier}
-• Разные результаты дают разные награды
-• 20 - критический успех!
-• 1 - критический провал!
-`;
-
-  await ctx.reply(message, {
-    parse_mode: "Markdown",
-    reply_markup: {
-      inline_keyboard: [
-        [{ text: "🎲 Бросить кубик!", callback_data: "quest_roll" }],
-      ],
-    },
-  });
-}
+  /
 
 // История квестов
 async function handleListQuests(ctx) {
@@ -3701,169 +3275,9 @@ async function handleDebugCharacters(ctx) {
 }
 
 // Прямой ввод имени через команду
-async function handleSetName(ctx) {
-  const userId = ctx.from.id;
-  const chatId = ctx.chat.id;
-  const sessionKey = `${userId}_${chatId}`;
 
-  // Получаем сессию
-  const session = characterCreation.creationSessions.get(sessionKey);
 
-  if (!session || session.step !== "name") {
-    await ctx.reply(
-      `❌ Нет активной сессии создания персонажа на этапе ввода имени.\n\n` +
-        `Используйте /create для создания персонажа.`,
-      { parse_mode: "Markdown" }
-    );
-    return;
-  }
-
-  // Получаем имя из команды
-  const text = ctx.message.text;
-  const parts = text.split(" ");
-
-  if (parts.length < 2) {
-    await ctx.reply(
-      `❌ Использование: /setname ИмяПерсонажа\n\n` +
-        `Например: /setname Горак Сильный`,
-      { parse_mode: "Markdown" }
-    );
-    return;
-  }
-
-  // Собираем имя из всех частей после команды
-  const name = parts.slice(1).join(" ").trim();
-
-  // Валидация имени
-  if (name.length < 2 || name.length > 20) {
-    await ctx.reply("❌ Имя должно быть от 2 до 20 символов!");
-    return;
-  }
-
-  if (!/^[а-яА-ЯёЁa-zA-Z\s-]+$/.test(name)) {
-    await ctx.reply("❌ Имя может содержать только буквы, пробелы и дефисы!");
-    return;
-  }
-
-  // Сохраняем имя и переходим к генерации характеристик
-  session.data.name = name;
-  session.step = "stats";
-
-  log(`[SetName] Имя установлено: "${name}" для сессии ${sessionKey}`);
-
-  await ctx.reply(
-    `✅ Имя принято: **${name}**\n\nГенерируем характеристики...`,
-    { parse_mode: "Markdown" }
-  );
-
-  // Запускаем генерацию характеристик (true = новое сообщение, без анимации)
-  await characterCreation.generateStats(ctx, true);
-}
-
-// Быстрое создание персонажа для групп
-async function handleQuickCreate(ctx) {
-  const text = ctx.message.text;
-  const parts = text.split(" ");
-
-  if (parts.length < 4) {
-    await ctx.reply(
-      `❌ **Использование:**\n` +
-        `/quickcreate раса класс имя\n\n` +
-        `**Расы:** human, elf, dwarf, halfling\n` +
-        `**Классы:** WARRIOR, ROGUE, MAGE, CLERIC, BARBARIAN, RANGER\n\n` +
-        `**Пример:**\n` +
-        `/quickcreate human WARRIOR Горак Сильный`,
-      { parse_mode: "Markdown" }
-    );
-    return;
-  }
-
-  const race = parts[1].toLowerCase();
-  const characterClass = parts[2].toUpperCase();
-  const name = parts.slice(3).join(" ").trim();
-
-  // Валидация
-  const config = require("../config/config");
-
-  if (!config.RACES[race]) {
-    await ctx.reply(`❌ Неверная раса! Доступны: human, elf, dwarf, halfling`);
-    return;
-  }
-
-  if (!config.CLASSES[characterClass]) {
-    await ctx.reply(
-      `❌ Неверный класс! Доступны: WARRIOR, ROGUE, MAGE, CLERIC, BARBARIAN, RANGER`
-    );
-    return;
-  }
-
-  if (name.length < 2 || name.length > 20) {
-    await ctx.reply("❌ Имя должно быть от 2 до 20 символов!");
-    return;
-  }
-
-  if (!/^[а-яА-ЯёЁa-zA-Z\s-]+$/.test(name)) {
-    await ctx.reply("❌ Имя может содержать только буквы, пробелы и дефисы!");
-    return;
-  }
-
-  try {
-    // Создаем/обновляем пользователя
-    const user = await User.findOrCreate(ctx.from);
-
-    // Проверяем, нет ли уже персонажа
-    const existing = await Character.findActive(ctx.from.id, ctx.chat.id);
-    if (existing) {
-      await ctx.reply(
-        `❌ У вас уже есть персонаж: **${existing.name}**!\n\n` +
-          `Используйте /delete для удаления.`,
-        { parse_mode: "Markdown" }
-      );
-      return;
-    }
-
-    await ctx.reply(
-      `🎲 **Создаем персонажа...**\n\n` +
-        `Раса: ${config.RACES[race].emoji} ${config.RACES[race].name}\n` +
-        `Класс: ${config.CLASSES[characterClass].emoji} ${config.CLASSES[characterClass].name}\n` +
-        `Имя: ${name}\n\n` +
-        `Генерируем характеристики...`,
-      { parse_mode: "Markdown" }
-    );
-
-    // Создаем персонажа с правильным user.id
-    const character = await Character.create(
-      user.id, // Используем ID из базы данных
-      ctx.chat.id,
-      name,
-      race,
-      characterClass
-    );
-
-    const display = await character.getFullDisplay();
-
-    await ctx.reply(
-      `🎉 **Персонаж создан!**\n\n${display}\n\n` +
-        `Используйте /hero для просмотра персонажа\n` +
-        `Квесты будут доступны с 10:00 до 22:00 МСК`,
-      { parse_mode: "Markdown" }
-    );
-
-    log(
-      `Быстрое создание персонажа: ${name} (${race} ${characterClass}) для пользователя ${ctx.from.id}`
-    );
-  } catch (error) {
-    log(`Ошибка быстрого создания персонажа: ${error.message}`, "error");
-
-    // Экранируем специальные символы для Markdown
-    const errorMessage = escapeMarkdown(error.message);
-
-    await ctx.reply(`❌ Ошибка создания: ${errorMessage}`, {
-      parse_mode: "Markdown",
-    });
-  }
-}
-
+    
 module.exports = {
   setupCommands,
 };
