+++ src/handlers/character/CharacterHandler.js
@@ -1,0 +1,1 @@
+//Создание и отображение персонажа.
