+++ src/systems/CharacterSystems/ClassSelector.js
@@ -1,0 +1,1 @@
+//Выбор класса персонажа.
