+++ src/systems/CharacterSystems/StatsGenerator.js
@@ -1,0 +1,1 @@
+// Генерация характеристик.
